+++ packages/av-canvas/vite.config.ts
@@ -15,9 +15,13 @@ export default defineConfig({
   },
   test: {
     browser: {
+      provider: 'webdriverio',
       enabled: true,
       name: 'chrome', // browser name is required
       headless: true,
+      providerOptions: {
+        browserVersion: 'stable',
+      },
     },
   },
   server: {