+++ packages/av-cliper/src/clips/iclip.ts
@@ -42,14 +42,14 @@ export interface IClip {
   clone: () => Promise<this>;
 
   /**
-   * 按指定时间切割，返回该时刻前后两个新素材
+   * 按指定时间切割，返回该时刻前后两个新素材，常用于剪辑场景按时间分割素材
    *
    * 该方法不会破坏原素材的数据
    *
    * @param time 时间，微秒
    * @returns
    */
-  split: (time: number) => Promise<[this, this]>;
+  split?: (time: number) => Promise<[this, this]>;
 
   /**
    * 销毁实例，释放资源