+++ packages/av-cliper/vite.config.ts
@@ -14,9 +14,13 @@ export default defineConfig({
   },
   test: {
     browser: {
+      provider: 'webdriverio',
       enabled: true,
       name: 'chrome', // browser name is required
       headless: true,
+      providerOptions: {
+        browserVersion: 'stable',
+      },
     },
   },
   server: {